requirements.txt

# API and web.
fastapi
uvicorn
jinja2
aiofiles
websockets
requests
# Audio, etc.
pyalsaaudio # For RPi.
### pyaudio # For macOS, etc.
numpy
scipy<1.9.0
psutil
pyusb
# Plotting.
matplotlib
watchdog
# GPS and date/time.
pyserial
pyserial-asyncio
pytz
python-dateutil
astral
# Config.
pyyaml
# Reports.
xlsxwriter
# VS Code.
pylint
black